Update .test_files method.
It now uses a webdav method to list out files in each study folder as a JSON. This list of files is compared to the list of files that are supposedly avaialable according to "getDataset" method.

Instead of using httr and jsonlite, it uses RCurl and rjson to minimize conflicts with other dependencie and to be consistent with Rlabkey.