Add interactive plot option in `quick_plot()` function.
For example, `study$quick_plot("hai", interactive = TRUE)` will show an interactive visualization on "Viewer" panel instead of showing a static image on "Plots" panel  in Rstudio. The default option is `FALSE`.

This will only apply to boxplot, violinplot, and lineplot, but not to heatmap since `plotly` doesn't have a function to convert a `pheatmap` object to a `plotly` object. This will be fixed using d3heatmap or other frameworks in future.